added links for user so they can access their account. users can access inputs through their wallets, so it filters inputs by a specific wallet

+++ app/src/Controller/TypeController.php
@@ -181,6 +181,12 @@ class TypeController extends AbstractController
      */
     public function delete(Request $request, Type $type): Response
     {
+        if ($type->getWallet()->count()) {
+            $this->addFlash('warning', 'message_type_contains_wallet');
+
+            return $this->redirectToRoute('type_index');
+        }
+
         $form = $this->createForm(FormType::class, $type, ['method' => 'DELETE']);
         $form->handleRequest($request);
